:spparkles: Renommage de Checklist
